requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
django-request-logging==0.7.5
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
